pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "FastAPI PoC"
authors = ["Brad Rogan <brad.rogan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.20.0"
fastapi = "^0"
python-multipart = "~0.0.5"
pydantic = { extras = ["email"], version = "^1.10.5" }
jinja2 = "3.0.3"
sqlalchemy = "^2.0.4"
alembic = "^1.9.4"
pylint-pydantic = "^0.1.6"
httpx = "^0.23.3"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.2"
pylint = "^2.16.2"

[tool.pylint.MASTER]
load-plugins = "pylint_pydantic"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
